refactor(server): split server package into multiple files
